+++ SpatialGDK/Source/SpatialGDK/Private/Interop/WellKnownEntitySystem.cpp
@@ -8,13 +8,15 @@ DEFINE_LOG_CATEGORY(LogWellKnownEntitySystem);
 
 namespace SpatialGDK
 {
-WellKnownEntitySystem::WellKnownEntitySystem(const FSubView& SubView, USpatialWorkerConnection* InConnection, const int InNumberOfWorkers,
+WellKnownEntitySystem::WellKnownEntitySystem(const FSubView& SubView, USpatialWorkerConnection* InConnection,
+											 USpatialNetDriver* InNetDriver, const int InNumberOfWorkers,
 											 SpatialVirtualWorkerTranslator& InVirtualWorkerTranslator,
 											 UGlobalStateManager& InGlobalStateManager)
 	: SubView(&SubView)
 	, VirtualWorkerTranslator(&InVirtualWorkerTranslator)
 	, GlobalStateManager(&InGlobalStateManager)
 	, Connection(InConnection)
+	, NetDriver(InNetDriver)
 	, NumberOfWorkers(InNumberOfWorkers)
 {
 }
@@ -152,7 +154,7 @@ void WellKnownEntitySystem::OnMapLoaded() const
 // for the TranslationManager, otherwise the manager will never be instantiated.
 void WellKnownEntitySystem::InitializeVirtualWorkerTranslationManager()
 {
-	VirtualWorkerTranslationManager = MakeUnique<SpatialVirtualWorkerTranslationManager>(Connection, VirtualWorkerTranslator);
+	VirtualWorkerTranslationManager = MakeUnique<SpatialVirtualWorkerTranslationManager>(Connection, NetDriver, VirtualWorkerTranslator);
 	VirtualWorkerTranslationManager->SetNumberOfVirtualWorkers(NumberOfWorkers);
 }