+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialNetDriver.h
@@ -80,14 +80,15 @@ class SpatialDebuggerSystem;
 class ActorSystem;
 class SpatialRPCService;
 class SpatialRoutingSystem;
+class FSpatialHandoverManager;
 class SpatialLoadBalanceEnforcer;
-class InterestFactory;
+class UnrealServerInterestFactory;
 class WellKnownEntitySystem;
 class ClientConnectionManager;
 class InitialOnlyFilter;
 class CrossServerRPCSender;
 class CrossServerRPCHandler;
-class SpatialStrategySystem;
+class FSpatialStrategySystem;
 class FSkeletonEntityCreationStep;
 } // namespace SpatialGDK
 
@@ -234,9 +235,10 @@ public:
 	FSpatialNetDriverServerRPC* ServerRPCs = nullptr;
 
 	TUniquePtr<SpatialGDK::SpatialRoutingSystem> RoutingSystem;
-	TUniquePtr<SpatialGDK::SpatialStrategySystem> StrategySystem;
+	TUniquePtr<SpatialGDK::FSpatialStrategySystem> StrategySystem;
 	TUniquePtr<SpatialGDK::SpatialLoadBalanceEnforcer> LoadBalanceEnforcer;
-	TUniquePtr<SpatialGDK::InterestFactory> InterestFactory;
+	TUniquePtr<SpatialGDK::FSpatialHandoverManager> HandoverManager;
+	TUniquePtr<SpatialGDK::UnrealServerInterestFactory> InterestFactory;
 	TUniquePtr<SpatialVirtualWorkerTranslator> VirtualWorkerTranslator;
 
 	TUniquePtr<SpatialGDK::FSkeletonEntityCreationStep> SkeletonEntityCreationStep;
@@ -291,7 +293,7 @@ private:
 	TUniquePtr<SpatialGDK::CrossServerRPCSender> CrossServerRPCSender;
 	TUniquePtr<SpatialGDK::CrossServerRPCHandler> CrossServerRPCHandler;
 
-	SpatialGDK::EntityQueryHandler QueryHandler;
+	SpatialGDK::FEntityQueryHandler QueryHandler;
 
 	TMap<Worker_EntityId_Key, USpatialActorChannel*> EntityToActorChannel;
 	TSet<Worker_EntityId_Key> DormantEntities;