+++ SpatialGDK/Source/SpatialGDK/Public/EngineClasses/SpatialVirtualWorkerTranslationManager.h
@@ -3,9 +3,7 @@
 #pragma once
 
 #include "EngineClasses/SpatialVirtualWorkerTranslator.h"
-#include "Interop/ClaimPartitionHandler.h"
-#include "Interop/CreateEntityHandler.h"
-#include "Interop/EntityQueryHandler.h"
+#include "Interop/SpatialCommandsHandler.h"
 #include "SpatialCommonTypes.h"
 #include "SpatialConstants.h"
 
@@ -43,7 +41,8 @@ public:
 		Worker_EntityId SimulatingWorkerSystemEntityId;
 	};
 
-	SpatialVirtualWorkerTranslationManager(SpatialOSWorkerInterface* InConnection, SpatialVirtualWorkerTranslator* InTranslator);
+	SpatialVirtualWorkerTranslationManager(SpatialOSWorkerInterface* InConnection, USpatialNetDriver* NetDriver,
+										   SpatialVirtualWorkerTranslator* InTranslator);
 
 	void SetNumberOfVirtualWorkers(const uint32 NumVirtualWorkers);
 
@@ -64,6 +63,7 @@ public:
 
 private:
 	SpatialOSWorkerInterface* Connection;
+	USpatialNetDriver* NetDriver;
 
 	TArray<VirtualWorkerId> VirtualWorkersToAssign;
 	TArray<PartitionInfo> Partitions;
@@ -72,9 +72,7 @@ private:
 
 	bool bWorkerEntityQueryInFlight;
 
-	SpatialGDK::CreateEntityHandler CreateEntityHandler;
-	SpatialGDK::ClaimPartitionHandler ClaimPartitionHandler;
-	SpatialGDK::EntityQueryHandler QueryHandler;
+	SpatialGDK::FCommandsHandler CommandsHandler;
 
 	// Serialization and deserialization of the mapping.
 	void WriteMappingToSchema(Schema_Object* Object) const;