+++ SpatialGDK/Source/SpatialGDK/Public/Schema/SpatialDebugging.h
@@ -14,7 +14,7 @@ namespace SpatialGDK
 // SpatialDebugger on clients but which would not normally be available to clients.
 struct SpatialDebugging : AbstractMutableComponent
 {
-	static const Worker_ComponentId ComponentId = SpatialConstants::SPATIAL_DEBUGGING_COMPONENT_ID;
+	static constexpr Worker_ComponentId ComponentId = SpatialConstants::SPATIAL_DEBUGGING_COMPONENT_ID;
 
 	SpatialDebugging()
 		: AuthoritativeVirtualWorkerId(SpatialConstants::INVALID_VIRTUAL_WORKER_ID)