+++ SpatialGDK/Source/SpatialGDK/Public/Utils/EntityFactory.h
@@ -20,6 +20,7 @@ namespace SpatialGDK
 {
 class InterestFactory;
 class SpatialRPCService;
+struct QueryConstraint;
 
 class SPATIALGDK_API EntityFactory
 {
@@ -39,9 +40,9 @@ public:
 										  TArray<FWorkerComponentUpdate>& OutComponentUpdates, uint32& OutBytesWritten);
 	static UnrealMetadata CreateMetadata(const AActor& InActor);
 
-	static TArray<FWorkerComponentData> CreatePartitionEntityComponents(const Worker_EntityId EntityId,
+	static TArray<FWorkerComponentData> CreatePartitionEntityComponents(const FString& PartitionName, const Worker_EntityId EntityId,
 																		const InterestFactory* InterestFactory,
-																		const UAbstractLBStrategy* LbStrategy,
+																		const SpatialGDK::QueryConstraint& LoadBalancingConstraint,
 																		VirtualWorkerId VirtualWorker, bool bDebugContexValid);
 
 	static inline bool IsClientAuthoritativeComponent(Worker_ComponentId ComponentId)